requirements.txt
numpy~=1.18.1
Pillow~=7.1.2
numba~=0.50.0